requirements.txt
aiohttp==3.7.2
aioredis==1.3.1
aiosignal==1.2.0
alembic==1.7.6
aniso8601==9.0.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.4.0
Authlib==0.15.5
black==22.1.0
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mistune==2.0.2
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==2.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==0.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.1.2
pytest-asyncio==0.12.0
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.1.0
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
URLObject==2.4.3
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
